// src/interface.ts
export interface IApiTemplate {
  id: string;
  name: string;
  createdon: string;
  createdby: string;
  notificationname: string;
  notificationemail: string;
  type: string;
  message: string;
  pages: number;
  pagedata: string;
  offsetboxes: boolean;
  redirecturl: string;
  score?: number;
}

export interface IComparingTemp {
  id: string;
  name: string;
  type: string;
  message: string;
  pages: number;
  pagedata: string[][];
  score?: number;
}
// ...
export interface IApiSigboxTemplate {
  //for templates
  left: string;
  top: string;
  width: string;
  height: string;
  esigntype: string;
  boxid: string;
  pagenumber: string;
  font: string;
  fontsize: string;
  lineheight: string;
  fontcolor: string;
  signsetid: string;
  fieldname: string;
  fieldvalue: string;
  fieldlabel: string;
  defaultvalue: string;
  depfield: string;
  depfieldvalue: string;
  depoperator: string;
  fieldrequired: string;
  checkedvalue: string;
  uncheckedvalue: string;
  role: string;
  signerid: number;
  page: number;
  xp: number;
  yp: number;
  hp: number;
  wp: number;
  type: number;
  text: string;
  checked: boolean;
  required: boolean;
}
// ...
export interface IElement {
  y: number;
  x: number;
  width: number;
  height: number;
  name: string;
  signer: string;
  options: IOptions[];
  id: string;
  type: string;
  required?: boolean;
  backgroundColor?: string;
  checkedvalue?: string;
  uncheckedvalue?: string;
  defaultvalue?: string;
}

export interface IOptions {
  label: string;
  value: string;
}